+++ app/src/main/AndroidManifest.xml
@@ -13,16 +13,19 @@
         android:theme="@style/Theme.Galaxy"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity2"
-            android:exported="false" />
-        <activity
-            android:name=".MainActivity"
+            android:name=".FrontPage"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
+        <activity android:name=".SignUpActivity"
+            android:exported="false">
+        </activity>
+        <activity android:name=".MainActivity"
+            android:exported="false">
         </activity>
     </application>
