+++ app/src/main/java/com/robox/galaxy/MainActivity.kt
@@ -1,88 +1,13 @@
 package com.robox.galaxy
 
 import android.os.Bundle
-import android.util.Log
-import android.view.View
-import android.widget.Toast
+import android.os.PersistableBundle
 import androidx.appcompat.app.AppCompatActivity
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.ktx.auth
-import com.google.firebase.ktx.Firebase
-import com.robox.galaxy.databinding.ActivityMainBinding
-import androidx.fragment.app.Fragment
 
-
-
-class MainActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityMainBinding
-    private lateinit var auth: FirebaseAuth
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        auth = Firebase.auth
-
-        binding.btnLogin.setOnClickListener {
-            loginUser()
-        }
-
-        binding.btnSignUp.setOnClickListener {
-            signUpUser()
-        }
+class MainActivity : AppCompatActivity()
+{
+    override fun onCreate(savedInstanceState: Bundle?, persistentState: PersistableBundle?) {
+        super.onCreate(savedInstanceState, persistentState)
+        setContentView(R.layout.mainactivity)
     }
-
-    private fun loginUser() {
-        val email = binding.etEmail.text.toString()
-        val password = binding.etPassword.text.toString()
-
-        if (email.isNotEmpty() && password.isNotEmpty()) {
-            binding.progressBar.visibility = View.VISIBLE
-            auth.signInWithEmailAndPassword(email, password)
-                .addOnCompleteListener(this) { task ->
-                    binding.progressBar.visibility = View.GONE
-                    if (task.isSuccessful) {
-                        Log.d("Auth", "signInWithEmail:success")
-                        Toast.makeText(this, "Login successful!", Toast.LENGTH_SHORT).show()
-                    } else {
-                        Log.w("Auth", "signInWithEmail:failure", task.exception)
-                        Toast.makeText(this, "Login failed: ${task.exception?.message}", Toast.LENGTH_SHORT).show()
-                    }
-                }
-        } else {
-            Toast.makeText(this, "Email/Password cannot be empty", Toast.LENGTH_SHORT).show()
-        }
-    }
-
-    private fun signUpUser() {
-        val email = binding.etEmail.text.toString()
-        val password = binding.etPassword.text.toString()
-
-        if (email.isNotEmpty() && password.isNotEmpty()) {
-            binding.progressBar.visibility = View.VISIBLE
-            auth.createUserWithEmailAndPassword(email, password)
-                .addOnCompleteListener(this) { task ->
-                    binding.progressBar.visibility = View.GONE
-                    if (task.isSuccessful) {
-                        Log.d("Auth", "createUserWithEmail:success")
-                        Toast.makeText(this, "Account created!", Toast.LENGTH_SHORT).show()
-                    } else {
-                        Log.w("Auth", "createUserWithEmail:failure", task.exception)
-                        Toast.makeText(this, "Sign-up failed: ${task.exception?.message}", Toast.LENGTH_SHORT).show()
-                    }
-                }
-        } else {
-            Toast.makeText(this, "Email/Password cannot be empty", Toast.LENGTH_SHORT).show()
-        }
-    }
-
-    override fun onStart() {
-        super.onStart()
-        if (auth.currentUser != null) {
-            Toast.makeText(this, "Already logged in as ${auth.currentUser?.email}", Toast.LENGTH_SHORT).show()
-        }
-    }
-
-
 }
