+++ app/src/main/java/com/robox/galaxy/SignUpActivity.kt
@@ -1,0 +1,11 @@
+package com.robox.galaxy
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+
+class SignUpActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.signup)
+    }
+}
