--- app/src/main/res/layout/signuppage.xml
@@ -1,208 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-
-    <RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    android:fillViewport="true"
-    android:padding="10dp"
-    android:background="@color/white"
-        >
-
-    <TextView
-        android:id="@+id/textView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_marginStart="40dp"
-        android:text="Create new account!"
-        android:textSize="30dp"
-        android:textStyle="bold" />
-
-    <TextView
-        android:id="@+id/tvSignup"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-
-        android:layout_alignParentStart="true"
-        android:layout_alignParentBottom="true"
-        android:layout_marginStart="94dp"
-        android:layout_marginBottom="626dp"
-        android:paddingBottom="30dp"
-        android:text="Already have account ?signup"
-        android:textColor="@color/BLUE400" />
-
-    <!--        &lt;!&ndash; Name Field &ndash;&gt;-->
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="26dp"
-        android:layout_marginTop="146dp"
-        android:layout_marginBottom="4dp"
-        android:text="Name"
-        android:textSize="16sp"
-        android:textStyle="bold" />
-
-    <EditText
-        android:id="@+id/etName"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentEnd="true"
-        android:layout_marginStart="26dp"
-        android:layout_marginTop="177dp"
-        android:layout_marginEnd="26dp"
-        android:hint="Enter your name"
-        android:inputType="textPersonName" />
-
-    <!--        &lt;!&ndash; Surname Field &ndash;&gt;-->
-    <!-- diffrence is 31dp-->
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="26dp"
-        android:layout_marginTop="235dp"
-        android:layout_marginBottom="4dp"
-        android:text="Surname"
-        android:textSize="16sp"
-        android:textStyle="bold" />
-
-    <EditText
-        android:id="@+id/etSurname"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentEnd="true"
-        android:layout_marginStart="26dp"
-        android:layout_marginTop="266dp"
-        android:layout_marginEnd="26dp"
-        android:hint="Enter your surname"
-        android:inputType="textPersonName" />
-
-    <!--        &lt;!&ndash; Enrollment Number Field &ndash;&gt;-->
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="26dp"
-        android:layout_marginEnd="26dp"
-
-        android:layout_marginTop="324dp"
-        android:layout_marginBottom="4dp"
-        android:text="Enrollment Number"
-        android:textSize="16sp"
-        android:textStyle="bold" />
-
-    <EditText
-        android:id="@+id/etEnrollment"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentEnd="true"
-        android:layout_marginStart="26dp"
-        android:layout_marginTop="355dp"
-        android:layout_marginEnd="26dp"
-        android:hint="Enter enrollment number"
-        android:inputType="number" />
-
-    <!--        &lt;!&ndash; Email Field &ndash;&gt;-->
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_marginStart="26dp"
-        android:layout_marginEnd="26dp"
-        android:textStyle="bold"
-        android:layout_marginTop="413dp"
-        android:layout_marginBottom="4dp"
-        android:text="Email"
-        android:textSize="16sp" />
-
-    <EditText
-        android:id="@+id/etEmail"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentEnd="true"
-        android:layout_marginStart="22dp"
-        android:layout_marginTop="444dp"
-        android:layout_marginEnd="26dp"
-        android:hint="Enter your email"
-        android:inputType="textEmailAddress" />
-
-    <!--        &lt;!&ndash; Department Selection &ndash;&gt;-->
-    <TextView
-        android:layout_width="114dp"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_alignParentTop="true"
-        android:layout_alignParentEnd="true"
-        android:layout_marginStart="29dp"
-        android:layout_marginTop="502dp"
-        android:layout_marginEnd="277dp"
-        android:layout_marginBottom="4dp"
-        android:text="Department"
-        android:textSize="16sp"
-        android:textStyle="bold" />
-
-
-
-
-        <LinearLayout
-            android:id="@+id/llGroupSelection"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical"
-            android:visibility="gone"
-            android:layout_marginTop="16dp">
-
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="Select Group"
-                android:textSize="16sp"
-                android:layout_marginBottom="4dp"/>
-
-            <RadioGroup
-                android:id="@+id/rgGroup"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:orientation="vertical">
-
-                <RadioButton
-                    android:id="@+id/rbGroup1"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="Group 1"/>
-
-                <RadioButton
-                    android:id="@+id/rbGroup2"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:text="Group 2"/>
-            </RadioGroup>
-        </LinearLayout>
-
-        <!-- Submit Button -->
-<!--    <Button-->
-<!--        android:id="@+id/btnSubmit"-->
-<!--        android:layout_width="match_parent"-->
-<!--        android:layout_height="wrap_content"-->
-<!--        android:layout_marginTop="24dp"-->
-<!--        android:text="Submit" />-->
-</RelativeLayout>
